--- src/classes/DB.js
@@ -1,14 +1,0 @@
-import mongoose from 'mongoose';
-import {userModel} from '../models/actionModel.js'
-
-export class DB {
-  
-    static init (cb) {
-        mongoose.connect(process.env.MONGO_URL, { useNewUrlParser: true }).then(()=>cb());
-    }
-
-    static async getFlows(userID){
-        const user = await userModel.findById(userID);
-        return user.flows;
-    }
-}
