--- src/routes/Actions/main.js
@@ -1,8 +1,0 @@
-import { Router } from "express"
-
-
-Router.get('/', (req, res) => {
-  res.send("Hello World!");
-})
-
-export const actionRouter= Router;
