+++ src/routes/webpanelHandler.js
@@ -1,14 +1,17 @@
 import express from "express";
-import { OAuth2Client } from "google-auth-library";
+import Auth from "../classes/auth";
 
 const Router = express.Router();
 
 Router.post('/', (req, res) => {
   const webpanelObj = req.body;
-  
+  let auth = new Auth;
   verify(webpanelObj.token).then((token) => {
     console.log(token);
     res.status(200).send('ok'); 
+
+
+
   }, (error) => {
     console.log("Failed authenticating: " + error.message);
     res.status(400).send(error.message); 
@@ -18,15 +21,3 @@ Router.post('/', (req, res) => {
   })
 
 export const webpanelHandler = Router;
-
-const client = new OAuth2Client(process.env.GOOGLE_TOKEN);
-
-async function verify(token) {
-  const ticket = await client.verifyIdToken({
-      idToken: token,
-      audience: process.env.GOOGLE_TOKEN,  // Specify the CLIENT_ID of the app that accesses the backend
-  });
-  const payload = ticket.getPayload();
-  const userid = payload.email;
-  return userid;
-}
